% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeScore.R
\name{cancers}
\alias{cancers}
\title{cancers Function}
\usage{
cancers()
}
\description{
This function returns a list of the currently supported cancers for the computeSignatureScore
}
\examples{
cancers()
}
\keyword{score}
