% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeScore.R
\name{computeSignatureScore}
\alias{computeSignatureScore}
\title{computeSignatureScore Function}
\usage{
computeSignatureScore(X, cancer)
}
\arguments{
\item{X}{gene dataframe expression dataset either in z-score or log2(n+1) normalized}

\item{cancer}{string that shows the names of the cancer}
}
\description{
This function computes the oncology models fidelity score based on the hallmarks of cancer
}
\examples{
cancers()
exampleOutput = computeSignatureScore(exampleInput, "Brain")
}
\keyword{score}
